+++ src/idtrackerai/base/postprocess/__init__.py
@@ -1,6 +1,8 @@
 from .get_trajectories import produce_output_dict
-from .trajectories_creation import trajectories_API
-from .trajectories_to_csv import convert_trajectories_file_to_csv_and_json
+from .trajectories_creation import (
+    convert_trajectories_file_to_csv_and_json,
+    trajectories_API,
+)
 
 __all__ = [
     "trajectories_API",