+++ src/idtrackerai/validator/validation_GUI.py
@@ -977,7 +977,7 @@ class SaveTrajectoriesThread(QThread):
         )
         if self.abort:
             return
-        trajectories_file = self.session.trajectories_folder / "validated.npy"
+        trajectories_file = self.session.trajectories_folder / "trajectories.npy"
         logging.info("Saving trajectories at %s", trajectories_file)
         np.save(trajectories_file, trajectories)  # type: ignore